requirements.txt
scrapy==2.11.2
kafka-python==2.0.2
python-dotenv==1.0.1
